requirements.txt
pandas
faker
duckdb
load_dotenv